Refactor MethodPresenter into own file
